+++ tests/dummy/app/locales/en/models/ember-flexberry-dummy-application-user.js
@@ -1,0 +1,24 @@
+export default {
+  projections: {
+    ApplicationUserL: {
+      name: {
+        __caption__: 'Name'
+      },
+      eMail: {
+        __caption__: 'E-mail'
+      },
+      activated: {
+        __caption__: 'Activated'
+      },
+      birthday: {
+        __caption__: 'Birthday'
+      },
+      gender: {
+        __caption__: 'Gender'
+      },
+      karma: {
+        __caption__: 'Karma'
+      },
+    },
+  }
+};